+++ extensions/fastmail-masked-email/src/create.tsx
@@ -33,9 +33,11 @@ export default function Command() {
     try {
       const email = await createMaskedEmail(prefix, description);
 
-      setMaskedEmail(email);
+      Clipboard.copy(email);
 
-      toast.hide();
+      await showHUD(`🎉 Masked email copied to clipboard`);
+
+      await closeMainWindow({ clearRootSearch: true, popToRootType: PopToRootType.Immediate });
     } catch (error) {
       toast.style = Toast.Style.Failure;
       toast.title = "Failed to create masked email";
@@ -46,21 +48,27 @@ export default function Command() {
     }
   };
 
-  const handleCopy = async () => {
-    if (!maskedEmail) {
-      await showToast({
-        style: Toast.Style.Failure,
-        title: "Copy failed",
-        message: "Create a masked email before attempting to copy",
-      });
+  const handleSubmitInteractive = async ({ prefix, description }: FormValues) => {
+    const toast = await showToast({ style: Toast.Style.Animated, title: "Creating masked email..." });
 
-      return;
+    try {
+      const email = await createMaskedEmail(prefix, description);
+
+      setMaskedEmail(email);
+
+      Clipboard.copy(email);
+
+      toast.style = Toast.Style.Success;
+      toast.title = "Masked email copied to clipboard";
+      toast.message = email;
+    } catch (error) {
+      toast.style = Toast.Style.Failure;
+      toast.title = "Failed to create masked email";
+      toast.message =
+        error instanceof Error
+          ? error.message
+          : "An error occurred while creating the masked email, please try again later";
     }
-
-    Clipboard.copy(maskedEmail);
-
-    await showHUD(`🎉 Masked email copied to clipboard`);
-    await closeMainWindow({ clearRootSearch: true, popToRootType: PopToRootType.Immediate });
   };
 
   return (
@@ -68,7 +76,11 @@ export default function Command() {
       actions={
         <ActionPanel>
           <Action.SubmitForm title="Create Masked Email" icon={Icon.EyeDisabled} onSubmit={handleSubmit} />
-          <Action.SubmitForm title="Copy Masked Email" icon={Icon.Clipboard} onSubmit={handleCopy} />
+          <Action.SubmitForm
+            title="Create Masked Email Interactively"
+            icon={Icon.EyeDisabled}
+            onSubmit={handleSubmitInteractive}
+          />
         </ActionPanel>
       }
     >
@@ -93,7 +105,7 @@ A prefix must be <= 64 characters in length and only contain characters a-z, 0-9
           <Form.Description
             text={`This masked email is currently in the pending state and will be automatically deleted after 24 hours if not used`}
           />
-          <Form.Description text={`Use ⇧⌘⏎ to copy the masked email to your clipboard`} />
+          <Form.Description text={`Use ⇧⌘⏎ to generate a new masked email`} />
         </>
       )}
     </Form>